# expense-tracker-service/expense_openapi.yaml
openapi: 3.0.0
info:
  title: Expense Tracker Service API
  description: API for managing Expenses.
  version: "1.0"
paths:
  /health:
    get:
      summary: Health Check
      responses:
        '200':
          description: Uptime
  /expensesByEmail/{id}:
    get:
      summary: Get all expenses by email
      parameters:
        - in: path
          name: email
          required: true
          schema:
            type: string
            format: email
          description: Filter by user email
      responses:
        '200':
          description: A list of expenses of given user
  /expenses/{id}:
    get:
      summary: Get a single expense by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
          description: Unique identifier of the expense
      responses:
        '200':
          description: Details of an expense
    delete:
      summary: Delete an expense
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
          description: Unique identifier of the expense to delete
      responses:
        '200':
          description: The expense was successfully deleted
  /expenses:
    get:
      summary: List all expenses by email
      parameters:
        - in: query
          name: email
          required: true
          schema:
            type: string
          description: Filter by user email
      responses:
        '200':
          description: A list of expenses of given user
    post:
      summary: Create a new expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Expense'
      responses:
        '201':
          description: The expense was successfully recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Expense'
components:
  schemas:
    Expense:
      type: object
      properties:
        amount:
          type: number
          format: double
        expenseGroup:
          type: string
        note:
          type: string
        expenseDate:
          type: string
          format: date
        email:
          type: string
          format: email
      required:
        - amount
        - expenseGroup
        - note
        - expenseDate
        - email